requirements.txt
Django==4.2.13
django-cors-headers==4.3.1
django-mongoengine==0.5.6
djangorestframework==3.15.1